Title: 특정 문자로 시작하는 문자열, Time: 117ms, Memory: 10MB, Status: Wrong Answer - CodeTree

+++ 240613/특정 문자로 시작하는 문자열/strings-that-start-with-a-specific-character.java
@@ -20,7 +20,7 @@ public class Main {
       }
       sum += arr[i].length();
     }
-    double aver = sum / n;
+    double aver = (double)sum / n;
     System.out.printf("%d %.2f", cnt, aver);
   }
 }
